backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2
Pillow
django-cors-headers>=3.13.0